// src/app/Models/Clientes/Juridicos/AccionistaModel.ts
export class AccionistaModel {
    IdJuridicoAccionista: number = 0;
    IdTercero: number = 0;
    IdTipoIdentificacion: number = 0;
    NumeroDocumento: string = "";
    RazonSocial: string = "";
    VinculoPeps: boolean = false;
    Participacion: number = 0;
    IdPersonaAccionista: number = 0;
}